requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.10.0
accelerate>=0.20.0
huggingface-hub>=0.15.0
python-dotenv>=1.0.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
jupyter>=1.0.0
tqdm>=4.65.0
peft>=0.4.0
bitsandbytes>=0.39.0 